pageing fix method corrected
